Extract src-tauri/src/telemetry.rs — both ends of the core mechanism
write_instrument_settings and run_telemetry_server go in one module because they
are one contract: the settings file generates the curl commands, the server
receives them, and a field renamed on either side goes quiet rather than
failing. That is what the Phase-0 server test exists to catch, and it moves here
with them (4 of the 69 tests: the instrument-settings assertions, query_param,
and the two that drive the real tiny_http server against a mock app).

resolve_permission came too, though it is a plain State<AppState> command like
a dozen others in lib.rs. It is the only thing that ever answers a held-open
PermissionRequest, so the blocking half of the mechanism is incomplete without
it — and the permission test calls it directly.

lib.rs 3,026 -> 2,595. Runtime dropped from its tauri imports: the generic
`run_telemetry_server<R: Runtime>`, added in Phase 0 so the tests could drive it
against tauri::test::mock_app(), was its only user.

One small finding while wiring: `use std::io::Read` is NOT needed by
run_telemetry_server despite the `request.as_reader().read_to_string(...)` call
— rustc flags it unused in the lib build. Dropped from module scope and added to
the test mod, which does need it for the raw-socket reads.

Verification:
- telemetry.rs vs the pre-slice lib.rs, line-sorted: only 3 differences, all
  test-mod imports that used to come from lib.rs's own test header (COUNTER,
  Read+Write, Ordering).
- the remainder of lib.rs: 9 lines, each intended — mod telemetry;, the use
  line, Runtime dropped, telemetry::resolve_permission in generate_handler!,
  and lib.rs's test mod no longer needing COUNTER/Ordering.

cargo test: 69 passed, same as baseline. cargo check --tests clean, no
warnings. No TypeScript touched.
